# api/openapi/q2-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "q2 API",
    "description": "Kudos (q2) — shared self-care goals. Reference API for the initial version.",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Creates an account and signs it in.",
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Signs in with an email address and a password.",
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/account": {
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Deletes the signed-in account and everything personal behind it. Asks for the password again.",
        "operationId": "DeleteAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/session": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Returns who is signed in.",
        "operationId": "GetSession",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/challenges/today/entry": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Contributes a photograph. A second one replaces the first.",
        "operationId": "SubmitChallengeEntry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitChallengeEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRoomResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Challenges"
        ],
        "summary": "Takes your contribution back. Your friends' are covered again afterwards.",
        "operationId": "WithdrawChallengeEntry",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeTodayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/challenges/entries/{id}/reactions": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Adds or takes back a reaction. The same kind twice takes it back.",
        "operationId": "ReactToChallengeEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactToChallengeEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeEntryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/feed/{id}/kudos": {
      "post": {
        "tags": [
          "Activity"
        ],
        "summary": "Gives kudos for an activity, or takes them back.",
        "operationId": "ToggleKudos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountDeletionResponse": {
        "required": [
          "goals",
          "images",
          "conversations",
          "messages",
          "challengeEntries"
        ],
        "type": "object",
        "properties": {
          "goals": {
            "type": "integer",
            "format": "int32"
          },
          "images": {
            "type": "integer",
            "format": "int32"
          },
          "conversations": {
            "type": "integer",
            "format": "int32"
          },
          "messages": {
            "type": "integer",
            "format": "int32"
          },
          "challengeEntries": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ActivityKind": {
        "enum": [
          "TaskCompleted",
          "StreakReached",
          "GoalProgress",
          "GoalCreated",
          "WindowAtRisk"
        ]
      },
      "ActivityResponse": {
        "required": [
          "id",
          "actor",
          "kind",
          "subject",
          "amount",
          "kudosCount",
          "hasMyKudos",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "actor": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "kind": {
            "$ref": "#/components/schemas/ActivityKind"
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kudosCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasMyKudos": {
            "type": "boolean"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ChallengeEntryResponse": {
        "required": [
          "id",
          "author",
          "imageId",
          "capturedInApp",
          "createdAt",
          "isMine",
          "reactions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "author": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "imageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "capturedInApp": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isMine": {
            "type": "boolean"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionSummaryResponse"
            }
          }
        }
      },
      "ChallengeResponse": {
        "required": [
          "id",
          "prompt",
          "publishedAt",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "prompt": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChallengeRoomResponse": {
        "required": [
          "challenge",
          "ownEntry",
          "entries",
          "friendCount",
          "isRevealed"
        ],
        "type": "object",
        "properties": {
          "challenge": {
            "$ref": "#/components/schemas/ChallengeResponse"
          },
          "ownEntry": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/ChallengeEntryResponse"
              }
            ]
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChallengeEntryResponse"
            }
          },
          "friendCount": {
            "type": "integer",
            "format": "int32"
          },
          "isRevealed": {
            "type": "boolean"
          }
        }
      },
      "ChallengeTodayResponse": {
        "required": [
          "room"
        ],
        "type": "object",
        "properties": {
          "room": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/ChallengeRoomResponse"
              }
            ]
          }
        }
      },
...
      "DeleteAccountRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "KudosKind": {
        "enum": [
          "Fire",
          "Strong",
          "Applause"
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PersonSummary": {
        "required": [
          "id",
          "displayName",
          "handle",
          "initials",
          "avatarColor",
          "isOnline",
          "avatarImageId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "avatarColor": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          },
          "avatarImageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ReactToChallengeEntryRequest": {
        "type": "object",
        "properties": {
          "kind": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/KudosKind"
              }
            ]
          }
        }
      },
...
      "ReactionSummaryResponse": {
        "required": [
          "kind",
          "count",
          "isMine"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/KudosKind"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "isMine": {
            "type": "boolean"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "inviteCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SessionResponse": {
        "required": [
          "person",
          "email"
        ],
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "SubmitChallengeEntryRequest": {
        "type": "object",
        "properties": {
          "imageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "capturedInApp": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  },
...
}